requirements.txt
django
django-rest-swagger
djangorestframework-jwt
drf-yasg

# aiohttp==3.8.4
# aiosignal==1.3.1
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# asgiref==3.6.0
# async-timeout==4.0.2
# attrs==22.2.0
# certifi==2022.12.7
# cffi==1.15.1
# charset-normalizer==3.0.1
# class-registry==2.1.2
# click==8.1.3
# colorama==0.4.6
# coreapi==2.3.3
# coreschema==0.0.4
# cryptography==39.0.0
# defusedxml==0.7.1
# diff-match-patch==20200713
# distlib==0.3.6
# Django==4.1.7
# django-allauth==0.52.0
# django-filter==22.1
# django-import-export==3.1.0
# django-multi-email-field==0.6.2
# django-rest-swagger==2.2.0
# djangorestframework==3.14.0
# djangorestframework-jwt==1.11.0
# djangorestframework-simplejwt==5.2.2
# djangorestframework-xml==2.0.0
# djongo==1.3.6
# dnspython==2.3.0
# drf-yasg==1.21.4
# et-xmlfile==1.1.0
# filelock==3.8.2
# filters==1.3.2
# Flask==2.2.3
# frozenlist==1.3.3
# geoip2==4.6.0
# geos==0.2.3
# googlemaps==4.10.0
# httpie==3.2.1
# idna==3.4
# inflection==0.5.1
# install==1.3.5
# itsdangerous==2.1.2
# itypes==1.2.0
# Jinja2==3.1.2
# lxml==4.9.2
# markdown-it-py==2.2.0
# MarkupPy==1.14
# MarkupSafe==2.1.2
# maxminddb==2.2.0
# mdurl==0.1.2
# multidict==6.0.4
# mysql==0.0.3
# mysqlclient==2.1.1
# numpy==1.24.2
# oauthlib==3.2.2
# odfpy==1.4.1
# openapi-codec==1.3.2
# openpyxl==3.1.2
# packaging==23.0
# pandas==1.5.3
# Pillow==9.3.0
# platformdirs==2.6.0
# pycparser==2.21
# pycryptodome==3.17
# pycryptodomex==3.10.1
# Pygments==2.13.0
# PyJWT==1.7.1
# pymongo==3.12.3
# PySocks==1.7.1
# python-dateutil==2.8.2
# python3-openid==3.2.0
# virtualenv==20.17.1
# virtualenvwrapper-win==1.2.7
# Werkzeug==2.2.3
# xlrd==2.0.1
# xlwt==1.3.0
# yarl==1.8.2